test(manager): test an sctool backup during enospc
In the test we run a backup on a node, stop it midway, fill its disk to 100% during the backup,
and run another backup afterwards. The second backup should run with no issue running
since the first backup already created snapshots that occupy some of the disk space,
and as such the second backup should go smoothly.