fix(manager-intensity): Added a second, smaller keyspace to be repaired in each node
Instead of deleting all files of keyspace1 in a single node and repairing it alone,
I've added a second keyspace and timed the insertion like so that each node
will have rows that will need to be repaired, making all of them equally in need of
a repair.

**node#1 is excluded due to an SCT limitation in the run_stress_thread function,
which demands that the node will be UN.

Also, enlarged the number of db nodes to 9.