feature(manager): sanities trigger upgrade and downstream jobs
To prevent triggering the manager upgrade and feature jobs unnecessarily,
I modified the manager pipeline as such that each of the jobs can trigger
another once they complete a successful run.
By default, each sanity triggers the matching upgrade job,
and the ipv6 sanity and backup feature jobs are
triggered by the centos sanity.
The user can edit the downstream_jobs_to_run param to choose
what jobs to trigger upon completion, but each job can only trigger jobs
that use the same lunux distribution.

Also, it should be noted that the sanity jobs should not be used to trigger
any other job, due to the nature of the pipeline.

At last, I'd like to specify that I placed the 'Running Downstream Jobs'
specifically after the test stage, so that even if the triggering job
fails on a stage like the 'Send email' stage, which is irrelevant to the
test result, the downstream jobs would still be executed.