fix(disrupt_nodetool_refresh): download the sstable tar to the runner
In some cloud runs, the nodes have no public IP, and as such cannot download files from s3,
like the sstable tarball that is downloaded in the disrupt_nodetool_refresh nemesis.
To amend that, I changed SstableLoadUtils.upload_sstables as such that in cloud runs,
the tarball would be first downloaded to the sct runner, and then moved to the node.