fix(compare_perftune_results): enveloped the function in a try clause
Since the compare_perftune_results function is being called in a very critical stage of the tests
(node_setup), any exception that will be raised by it can fail an entire run. So I encapsulated
most of the function in a try clause, and any exception caught by it will be published as an error
event.