fix(disrupt_mgmt_restore): remove quotation marks from keyspace names
In the beginning of the nemesis, we request a list of the existing keyspaces,
to see if the keyspace we wish to restore already exists.
That check was faulty, since it did not account for the fact that keyspaces
that start with a digit are surrounded by quotation marks.
I amended the check to account for that.