fix(longevity-5gb-1h-nemesis): set partition_range_with_data_validation
For delete nemeses (like disrupt_delete_overlapping_row_ranges) partition_range_with_data_validation
must be set, so the nemeses will know which partitions cannot be deleted.
Also changed the number of partitions that are validated from 10000 to 9990 so that there
will be any partitions that could be deleted by the nemeses in the first place.